Refactor master list line hooks
