+++ client/packages/system/src/MasterList/api/hooks/index.ts
@@ -1,12 +1,1 @@
-import { Lines } from './line';
-import { Utils } from './utils';
 export { useMasterLists } from './useMasterLists';
-
-export const useMasterListOld = {
-  line: {
-    rows: Lines.useMasterListLines,
-  },
-  utils: {
-    api: Utils.useMasterListApi,
-  },
-};
