--- client/packages/system/src/MasterList/api/hooks/line/index.ts
@@ -1,3 +1,0 @@
-import { useMasterListLines } from './useMasterListLines';
-
-export const Lines = { useMasterListLines };
