--- client/packages/system/src/MasterList/api/hooks/line/useMasterListLines.ts
@@ -1,20 +1,0 @@
-import {
-  useParams,
-  useQuery,
-  useUrlQueryParams,
-} from '@openmsupply-client/common';
-import { useMasterListApi } from '../utils/useMasterListApi';
-
-export const useMasterListLines = () => {
-  const { queryParams } = useUrlQueryParams({
-    initialSort: { key: 'itemName', dir: 'asc' },
-  });
-  const { id } = useParams();
-
-  const api = useMasterListApi();
-  const masterListId = id || '';
-
-  return useQuery(api.keys.lines(masterListId, queryParams), () =>
-    api.get.lines(masterListId, queryParams)
-  );
-};
