--- client/packages/system/src/MasterList/api/hooks/utils/index.ts
@@ -1,5 +1,0 @@
-import { useMasterListApi } from './useMasterListApi';
-
-export const Utils = {
-  useMasterListApi,
-};
